pyproject.toml
[tool.poetry]
name = "trade-machine"
version = "0.1.0"
description = ""
authors = ["Marcus Jefferson <marcjeffs6@gmail.com>"]
readme = "README.md"
packages = [{include = "trade_machine"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.100.0"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
alpaca-py = "^0.10.0"
sqlalchemy = "^2.0.20"
python-dotenv = "^1.0.0"

[tool.poetry.scripts]
trade-machine = "trade_machine.__main__:main"
dev = "trade_machine.__main__:dev"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
